requirements.txt
websocket-client==1.8.0
websockets==13.1
torchaudio==2.4.1
torch==2.4.1
torchvision==0.19.1
pyaudio==0.2.14
numpy==2.1.1
